pyproject.toml
[tool.poetry]
name = "catrius_blog_backend"
version = "0.1.0"
description = "Catrius's blog backend"
authors = ["Catrius Pham <catriuspham@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.0.2"
django-environ = "^0.4.5"
psycopg2 = "^2.8.4"
djangorestframework = "^3.11.0"
django-extensions = "^2.2.8"
django-cors-headers = "^3.2.1"
Pillow = "^7.0.0"
django-resized = "^0.3.9"
django-watson = "^1.5.4"
django-storages = "^1.9.1"
boto3 = "^1.12.17"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.3"
Faker = "^4.0.1"
factory_boy = "^2.12.0"
pytest-django = "^3.8.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
